Added support for int colors

+++ src/main.rs
@@ -14,9 +14,9 @@ pub mod util;
 pub struct Dot {
     pub x: f32,
     pub y: f32,
-    pub r: f32,
-    pub g: f32,
-    pub b: f32,
+    pub r: u8,
+    pub g: u8,
+    pub b: u8,
     pub size: f32,
 }
 
@@ -35,9 +35,9 @@ async fn retrieve(id: &str, pass: &str) -> Option<File> {
         let empty_vec: Vec<Dot> = vec![Dot {
             x: 0.0,
             y: 0.0,
-            r: 0.0,
-            g: 0.0,
-            b: 0.0,
+            r: 0,
+            g: 0,
+            b: 0,
             size: 0.0,
         }];
